+++ Sources/Tessera/Tessera.swift
@@ -1,2 +1,4 @@
 @_exported import TesseraCore
+@_exported import TesseraLayout
 @_exported import TesseraTerminal
+@_exported import TesseraWidgets